+++ SimpleRequestLogger.Benchmarks/Program.cs
@@ -1,0 +1,9 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace SimpleRequestLogger.Benchmarks
+{
+    public class Program
+    {
+        public static void Main() => BenchmarkRunner.Run<Benchmarks>();
+    }
+}