+++ SimpleRequestLogger/SimpleRequestLoggerMiddleware.cs
@@ -23,17 +23,22 @@ namespace SimpleRequestLogger
             ILoggerFactory loggerFactory, RequestDelegate next)
         {
             _Configuration = configuration;
-            _Configuration.MessageTemplate ??= "";
             _Logger = loggerFactory.CreateLogger(nameof(SimpleRequestLogger));
             _Next = next;
-            _PropertyNames = Regex.Matches(configuration.MessageTemplate, @"((?<=\{)[a-zA-Z]+(?=\}))")
+            _PropertyNames = Regex.Matches(_Configuration.MessageTemplate, @"((?<=\{)[a-zA-Z]+(?=\}))")
                 .Select(m => m.Value)
                 .ToList();
-            VerifyConfiguration();
+            _ = MapProperties(new LoggingContext(default, default, default));
         }
 
         public async Task InvokeAsync(HttpContext httpContext)
         {
+            if (_Configuration.IgnorePathPatterns.Any(p => Regex.IsMatch(httpContext.Request.Path, p)))
+            {
+                await _Next(httpContext);
+
+                return;
+            }
             var start = Stopwatch.GetTimestamp();
             var statusCode = 0;
             try
@@ -49,36 +54,12 @@ namespace SimpleRequestLogger
             }
             finally
             {
-                var elapsedMs = (Stopwatch.GetTimestamp() - start) * 1000 / (double)Stopwatch.Frequency;
+                var elapsedMs = (int)Math.Round((Stopwatch.GetTimestamp() - start) * 1000 / (double)Stopwatch.Frequency);
                 var logLevel = _Configuration.LogLevelSelector(statusCode);
                 var loggingContext = new LoggingContext(httpContext, statusCode, elapsedMs);
 #pragma warning disable CA2254 // Template should be a static expression.
                 _Logger.Log(logLevel, _Configuration.MessageTemplate, MapProperties(loggingContext));
 #pragma warning restore CA2254 // Template should be a static expression.
-            }
-        }
-
-        private void VerifyConfiguration()
-        {
-            if (Regex.IsMatch(_Configuration.MessageTemplate,
-                @"(\{[^\}]*\{)|(\}[^\{]*\})|((?<=\{)[a-zA-Z]*[^a-zA-Z\{\}]+[a-zA-Z]*(?=\}))|(\{\})"))
-            {
-                throw new InvalidOperationException("Message template is invalid.");
-            }
-            _ = MapProperties(new LoggingContext(default, default, default));
-            var statusCodes = typeof(StatusCodes).GetFields()
-                .Select(fi => (int)fi.GetValue(null)!)
-                .Distinct();
-            foreach (var statusCode in statusCodes)
-            {
-                try
-                {
-                    _ = _Configuration.LogLevelSelector(statusCode);
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException($"Log level selector throws an exception on status code {statusCode}.", e);
-                }
             }
         }