+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
 
+## [1.1.0] - 2022-02-22
+### Added
+- Ability to ignore specific paths
+- Performance benchmarks
+- XML documentation for all public types and members
+- More configuration validations
+### Changed
+- ElapsedMs type to integer
+- Default message template
+
 ## [1.0.0] - 2022-01-12
 - Initial release